iOS: Split the UTI coordinator test file (part 3)
UnifiedToggleInputCoordinatorTests.swift held three test classes. Moves the
two that aren't coordinator tests into their own files, and extracts the one
mock two of them shared.

Pure relocation: the same 320 tests pass before and after, and the only
content change in the whole diff is dropping `private` from the shared mock.
The toolbar-view class turned out to hold eleven coordinator tests as well as
its seven view tests, so those eleven move to the coordinator suite rather
than into a view-test file.
